recipes/ESC50/interpret/extra_requirements.txt
matplotlib
pandas
scikit-learn
torchvision
transformers
